Agent: Move PortScanData to its own module
