+++ monkey/infection_monkey/i_puppet/__init__.py
@@ -1,7 +1,7 @@
+from .port_scan_data import PortScanData
 from .i_puppet import (
     IPuppet,
     ExploiterResultData,
-    PortScanData,
     FingerprintData,
     UnknownPluginError,
 )
