+++ monkey/infection_monkey/i_puppet/port_scan_data.py
@@ -1,0 +1,13 @@
+from typing import Optional
+
+from pydantic import Field
+
+from common.base_models import InfectionMonkeyBaseModel
+from common.types import NetworkPort, PortStatus
+
+
+class PortScanData(InfectionMonkeyBaseModel):
+    port: NetworkPort
+    status: PortStatus
+    banner: Optional[str] = Field(default=None)
+    service: Optional[str] = Field(default=None)
